pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kaizen-studio"
version = "0.1.0"
description = "Kaizen Studio - AI Agent Platform Backend"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Kaizen Team", email = "team@kaizen.dev"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Kailash SDK & Frameworks
    "kailash>=0.10.13",
    "kailash-nexus>=1.1.2",
    "kailash-dataflow>=0.10.5",
    "kailash-kaizen>=0.7.1",

    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Authentication
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",

    # Database & Cache
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "motor>=3.3.0",
    "psycopg2-binary>=2.9.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-dotenv>=0.5.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-dotenv>=0.5.2",
    "httpx>=0.25.0",
    "fakeredis>=2.20.0",
    "respx>=0.21.1",  # HTTP mocking for httpx
]

[project.urls]
Homepage = "https://github.com/kaizen/kaizen-studio"
Documentation = "https://docs.kaizen.dev"
Repository = "https://github.com/kaizen/kaizen-studio"

[tool.hatch.build.targets.wheel]
packages = ["src/studio"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
env_files = [".env.test"]
addopts = "-v --cov=studio --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from err (pre-existing, would require extensive refactoring)
    "E402",  # module level import not at top (DataFlow requires setup before imports)
    "F841",  # unused variable (pre-existing in codebase)
    "B007",  # loop control variable not used (pre-existing)
]

[tool.ruff.lint.per-file-ignores]
"src/studio/config/__init__.py" = ["F401"]  # Re-exports for package interface

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/studio"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
